+++ src/controllers/reportsController.js
@@ -6,8 +6,8 @@ const {
   serverTimestamp,
   addDoc,
 } = require("firebase/firestore");
-const db = require("../firebase");
-const { CRUD } = require("../module/crud");
+const db = require("../config/firebase");
+const { CRUD } = require("../utils/crud");
 
 const reportsRef = collection(db, "reports");
