+++ src/controllers/statusController.js
@@ -1,9 +1,9 @@
 const {
   collection,
 } = require("firebase/firestore");
-const db = require("../firebase");
-const { CRUD } = require("../module/crud");
-const { getLast30Days, getLast7Days, getLatest } = require("../module/getLatest");
+const db = require("../config/firebase");
+const { CRUD } = require("../utils/crud");
+const { getLast30Days, getLast7Days, getLatest } = require("../utils/getLatest");
 
 const statusRef = collection(db, "status");
 const yearlyRef = collection(db, 'yearly')
