+++ src/routes/wasteRouter.js
@@ -1,10 +1,10 @@
 const express = require('express')
 const { WasteController } = require('../controllers/wasteController')
 const { collection } = require('firebase/firestore')
-const db = require('../firebase')
-const { calculateTotalMiddleware } = require('../middlewares/calculateTotal')
-const { calculateAverageMiddleware } = require('../middlewares/calculateAverage')
-const { calculateMonthlyMiddleware } = require('../middlewares/calculateMonthly')
+const db = require('../config/firebase')
+const { calculateTotalMiddleware } = require('../services/calculateTotal')
+const { calculateAverageMiddleware } = require('../services/calculateAverage')
+const { calculateMonthlyMiddleware } = require('../services/calculateMonthly')
 const router = express.Router()
 
 const wasteRef = collection(db, 'waste')
